refactored app for structure to match tests
