+++ public/index.html
@@ -2,6 +2,7 @@
 <html>
   <head>
     <title>Stopwatch</title>
+    <link rel="icon" href="/public/favicon.ico" />
   </head>
   <body>
     <section id="root"></section>
