+++ src/App.css
@@ -26,6 +26,7 @@ h1 {
 .btn-container {
   display: flex;
   gap: 1rem;
+  margin-bottom: 2rem;
 }
 
 button {
